use harmony priority, remove 583429740 TMPE support
